RevendaController adicionado
